backend/requirements.txt
fastapi
uvicorn
cryptography
python-dotenv
web3
pydantic
